Centralized randomness generators into one price-enforcement script which runs all at once as async functions and places orders at random intervals. Also allowed shorting up to 1000 shares. Had tons of problems changing variable types, turns out issue was being cause by build.rs

+++ src/config.rs
@@ -101,11 +101,11 @@ macro_rules! generate_account_balances_struct {
     ([$($name:ident),*]) => {
         #[derive(Debug, Serialize, Deserialize)]
         pub struct AssetBalances {
-            $($name: Mutex<usize>, )*
+            $($name: Mutex<i64>, )*
         }    
 
         impl AssetBalances {
-            pub fn index_ref (&self, symbol:&TickerSymbol) -> &Mutex<usize>{
+            pub fn index_ref (&self, symbol:&TickerSymbol) -> &Mutex<i64>{
                 match symbol {
                     $(TickerSymbol::$name => {&self.$name}, )*
                 }
